added committee subscription component, loading overlay moved to app.vue
